Extract `TypedArray` typings to a separate file
