+++ cupLayoutTool/src/jsMain/kotlin/three/three.kt
@@ -358,39 +358,6 @@ fun LineBasicMaterialParams(
     return obj
 }
 
-abstract external class TypedArray<T> {
-    val length: Int
-
-    fun set(array: Array<T>, offset: Int = definedExternally)
-}
-
-// TypedArray definitions needed for Three.js
-@JsName("Float32Array")
-external class Float32Array(array: Array<Double>) : TypedArray<Double> {
-    constructor(size: Int)
-}
-
-@Suppress("NOTHING_TO_INLINE")
-inline operator fun Float32Array.get(index: Int): Double = asDynamic()[index]
-
-@Suppress("NOTHING_TO_INLINE")
-inline operator fun Float32Array.set(index: Int, newItem: Double) {
-    asDynamic()[index] = newItem
-}
-
-@JsName("Uint16Array")
-external class Uint16Array(array: Array<Int>) : TypedArray<Int> {
-    constructor(size: Int)
-}
-
-@Suppress("NOTHING_TO_INLINE")
-inline operator fun Uint16Array.get(index: Int): Int = asDynamic()[index]
-
-@Suppress("NOTHING_TO_INLINE")
-inline operator fun Uint16Array.set(index: Int, newItem: Int) {
-    asDynamic()[index] = newItem
-}
-
 fun THREE.Vector3.toList(): List<Double> = listOf(
     this.x,
     this.y,
